moved stories in the right folder
